Async find and where methods
In order to implement this, I had to make some breaking changes. I'm
also changing how the Registry keeps track of things.

The registry now works off of JSON-API types - this makes translation /
mapping possible when the data is coming in.

When we don't have a registered model, we throw a `console.warn` and use
the base model.

The tests have closer coupling than I would like. I'm open to
refactoring them so they're less coupled and more single-purpose, but I
wanted to finish this up first.