Add functionality to update reverse relationship
Closes #9

When a resource is merged, it's possible that it will have one-sided
relationships. We're a good friend of this resource, and we think it
deserves to be happy, so we're going to inform the other parties of
their relationship status.

Example:

```js
/*
 * We now have Person number 5 in our store. Like Mourinho after a tough
 * match, it has no comments.
 */
Person.find(5);

/*
 * When we create a comment for the person above, the information we get
 * back will indicate that the comment belongs to the person
 */
dispatch(createComment(data));

/*
 * When we query the store now, person number 5 has comments.
 * There is much rejoicing.
 */
Person.find(5)
```

Also:
Rename hasProperty to getProperty