Extract queries in preparation for new Connect
The current connect component feels flawed, unstable and overly complex.
Rather than having our Connect component need to resolve promises, we
can take advantage of Redux + Reselect to hold and index results from
promises, which can then dynamically update data fairly inexpensively.

The idea of indexing records came up as I was attempting to solve
pagination in an external plugin. As I tried to solve for how we would
keep the data up-to-date based on information flowing into the store, I
realized that the connect component simply has to do too much. It's not
performant, and it wouldn't necessarily always re-render when we need it
to.