Implement index reducer
The index reducer changes how we interact with our resources. We now
have a place to wait for our queries while they are fetched. As a
result, we do not need to have our component interact with promises.
Rather, we can point our component at an empty array, which is mapped
over resources in the store upon change.

**Before**

1. `Person.find(id)`
2. Make request to server
3. Get answer from server
4. Send data into store
5. Instantiate objects off of promise

What you can see here is that at no point can we lose our reference
to this promise. We are dependent on it to tell us what resources we
are serving for a page.

**After**

1. `Person.find(id)`
2. Create index with hash based on query (`people.id=[id]`)
3. Make request to server
4. Send data to resources store, add new resources into index
5. Reselect selector instantiates objects off of promise
(6. Unsubscribe / clear index when component is unmounted)

Now we have the places in our store to watch synchronously without
needing to care about promises and can rely on native `connect` /
`reselect` implementations to refresh our data.