Create extension architecture
This PR also contains significant cleanup and an overall improvement in
the quality of tests being run. The old tests were heavily borrowed from
redux-json-api - which, while it provided a good foundation to start on,
did not test large parts of the application and relied on re-creating
actions in order to test them.

The new tests only cover public API. As a result, they are more
comprehensive. They do some stubbing around axios in order to work.

Another thing that was changed is the consolidation of files for the
main reducer. We at CoverHound have been moving towards the "ducks" type
of reducer organization as it's easier to keep context if you don't have
to go through multiple files to make a reducer change.

This has the nice side effect of showcasing how an extension would work.