Fix publish so it includes dist
We can verify this is working by running `npm pack` in the local
directory. Including `src` just so that people running these
node_modules can have the original reference, though I'm also open to
the idea of removing that directory.

Added a script for publishing because we can't reliably know that
`prepublish` is going to only run on publish, and we still want to
ensure that people remember to run all the steps before publishing.