Introduce reducer namespace
Currently it is a _requirement_ to use 'api' as the object key in
`combineReducers`. This PR introduces a code change that should make the
key customizable. The big use-case here is if you have name conflicts,
or what to gradually introduce `active-redux` into your codebase.

This library has a tight coupling to the chosen namespace of
`combineReducers` due to various references to `state.api` in the
codebase. By having the namespace managed by a central module we can
keep with the default 'api' but also open it up to customizability.

**IMPORTANT**

This is a very experimental feature and _may be removed_. Because of
this, I'm being explicitly not documenting this feature. Use at your own
risk!